requirements.txt
# Core
numpy==1.26.4
psutil==5.9.8

# PyTorch (CPU-only or CUDA depending on your setup)
torch==2.8.0+cu126
torchvision==0.19.0+cu126
torchaudio==2.8.0+cu126
# ^ If you want CPU-only, use the "+cpu" wheels instead.

# Sentence Transformers
sentence-transformers==3.1.1

# FAISS (CPU or GPU version)
faiss-cpu==1.8.0.post1
# or faiss-gpu if you built with CUDA support

# PaddleOCR (and deps)
paddlepaddle==2.6.1
paddleocr==2.7.0

# PDF / Document parsing
PyMuPDF==1.24.10
pillow==10.4.0
beautifulsoup4==4.12.3
